massgrowth_sas/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
instagrapi==2.1.5
sqlalchemy==2.0.25
aiosqlite==0.19.0
apscheduler==3.10.4
pydantic-settings==2.1.0
pyyaml==6.0.1
structlog==24.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
